Refactored model file into separate modules, merged with dev, renamed EmailConfirmation to not use plural
