+++ src/registrationSystem/models/common/__init__.py
@@ -1,4 +1,3 @@
-from .abstract_user import AbstractBaseUser
-from .abstract_group import AbstractGroup
-from .abstract_singleton_model import AbstractSingletonModel
+# flake8: noqa
 from .common_user_manager import CommonUserManager
+from .important_date import ImportantDate
