+++ src/registrationSystem/models/common/abstract_group.py
@@ -1,5 +1,5 @@
-from django.db import models
 from django.contrib.auth import get_user_model
+from django.db import models
 
 
 class AbstractGroup(models.Model):
