+++ src/registrationSystem/models/common/important_date.py
@@ -1,0 +1,5 @@
+from django.db import models
+
+
+class ImportantDate(models.Model):
+    date = models.DateField()
