+++ src/registrationSystem/models/river_rafting/__init__.py
@@ -1,0 +1,13 @@
+# flake8: noqa
+from django.conf import settings
+from django.contrib import admin
+
+from .river_rafting_raffle_state import RiverRaftingRaffleState
+from .river_rafting_user import RiverRaftingUser
+from .river_rafting_team import RiverRaftingTeam
+from .interest_check import InterestCheck
+from .email_confirmation import EmailConfirmation
+
+if settings.EVENT == 'RIVERRAFTING':
+    admin.site.register(RiverRaftingUser)
+    admin.site.register(RiverRaftingTeam)
